requirements.txt
django>=2.2.4
psycopg2==2.8.3
djangorestframework==3.10.3
